backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.34.0
tortoise-orm==0.20.0
asyncpg==0.29.0
pydantic==2.6.1
pydantic[email]==2.6.1
werkzeug==3.0.1
watchdog==6.0.0
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
python-dotenv==1.0.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt<4.0
